Move youtube.helper.channel_filter_split method to kodion.utils.convert_format
- Also move youtube.helper.custom_filter_split used by channel_filter_split method
